Include compiler versions in build directory name
Before, the build script ignored differing versions of the C and
Fortran compilers; it built OpenBLAS in a directory named only by the
target. Now, the build script gets information about the build
environment, including the C and Fortran compiler versions, and
includes a hash of that information in the build directory name. It
also writes the build information to a file named
`.openblas-src_build_info` in the build directory to help with
debugging.

The result of this change is that when the C or Fortran compiler is
changed (e.g. by a system update or by the user specifying the
compiler version in environment variables), OpenBLAS will be rebuilt.

The hasher selected for hashing the build information is
`std::collections::hash_map::DefaultHasher::new()` simply because it's
included in the standard library, so the build script doesn't need
extra dependencies. The disadvantage of this hasher is that its output
is not guaranteed to be stable across different Rust releases. As a
result, with a new Rust release, OpenBLAS may be rebuilt
unnecessarily.