Print rustc-link-search for the in-tree static openblas before everything else
Fixes #60.

The gnu linker searches libraries in the order of -L options and stops
the search when the first matching library is found. This means if
libopenblas.a is either in deliv.make_conf.c_extra_libs.search_path or
in deliv.make_conf.f_extra_libs.search_path, the in-tree libopenblas.a
won't be used.

This patch fixes the issue by printing the rustc-link-search instruction
for the in-tree openblas before everything else.